Fixes based on update to average growth calc
After discussing with Horace the fact that the spec needed to be updated to specify that average growth is:
 "where 0.123 is their average percentage growth across years. If there are three years of proficiency data (year1, year2, year3), that's ((proficiency at year3) - (proficiency at year1)) / (year3 - year1)"
some of the calculations need to be updated now that the denominator is based solely on the distance between the last year of data and the first year of data. Basically if there are 'N/A' or nil values (whatever the group is using) inbetween the first and last year, the denominator used to be different (since the count of total years of data was less than the distance between the first and last years). Since only the first and last year matter now, 'N/A' values in the middle no longer factor into the numerator or denominator.